+++ images/base-python-adapter/swagger_server/controllers/controller.py
@@ -7,7 +7,7 @@ import logging
 import os.path
 import subprocess
 import tempfile
-import threading_temp
+import threading
 
 import connexion
